Overlaps
Write a program that reads the coordinates of four point points, the
former two defining one rectangle and the
latter two defining another one. The program must read the coordinates
of a fifth point and say whether the point
is inside both rectangles, inside one of them only, or out of both.